Move initializable event data
