+++ StardewArchipelago/Locations/CodeInjections/Initializers/ModCodeInjectionInitializer.cs
@@ -11,10 +11,13 @@ namespace StardewArchipelago.Locations.CodeInjections.Initializers
     public static class ModCodeInjectionInitializer
     {
         static ArchipelagoClient _archipelago;
+        private static CallableModData _conditionInitializer;
 
         public static void Initialize(IMonitor monitor, IModHelper modHelper, ArchipelagoClient archipelago, LocationChecker locationChecker, ShopReplacer shopReplacer, ShopStockGenerator shopStockGenerator)
         {
             _archipelago = archipelago;
+            _conditionInitializer = new CallableModData();
+            _conditionInitializer.Initialize(monitor, archipelago);
             InitializeModdedContent(monitor, modHelper, archipelago, locationChecker, shopReplacer, shopStockGenerator);
         }
